Create utils file and move add auth in there
