+++ lib/utils.js
@@ -1,0 +1,32 @@
+var basicAuth = require('basic-auth');
+
+/**
+ * Simple basic auth middleware for use with Express 4.x.
+ *
+ * Based on template found at: http://www.danielstjules.com/2014/08/03/basic-auth-with-express-4/
+ *
+ * @example
+ * app.use('/api-requiring-auth', utils.basicAuth('username', 'password'));
+ *
+ * @param   {string}   username Expected username
+ * @param   {string}   password Expected password
+ * @returns {function} Express 4 middleware requiring the given credentials
+ */
+exports.basicAuth = function(username, password) {
+	return function(req, res, next) {
+
+		if (!username || !password) {
+			console.log('Username or password is not set.');
+			return res.send('Error: username or password not set. <a href="https://github.com/alphagov/govuk_prototype_kit/blob/master/docs/deploying.md#3-set-a-username-and-password">See guidance for setting these</a>');
+		}
+		
+		var user = basicAuth(req);
+
+		if (!user || user.name !== username || user.pass !== password) {
+			res.set('WWW-Authenticate', 'Basic realm=Authorization Required');
+			return res.send(401);
+		}
+
+		next();
+	};
+};
